+++ mithril-stm/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.9.17 (02-19-2026)
+
+### Added
+
+- Added the computation of the lottery target value for SNARK.
+- Switched back-end for computation from rug to num-integer.
+
 ## 0.9.16 (02-18-2026)
 
 ### Changed